Added database manager flask/sqlite3 program
uses Flask, jsonify, flask_cors, sqlite3 (sqlite3 is built-in with python).  Creates a database called devices.db on startup (if it doesn't already exist) and adds two test devices to the database.  Devices are located in the 'devices' table and have the attributes id, mac, ip, product, vendor, type, status, notes.  It's a whole lot easier to type in an id correctly than it is to type in a mac address correctly for deleting/editing a device, and it also allows users to edit the mac of a device if it was typed in wrong.  Currently, there is no error checking, all actions commit to the database, and post requests are not set up properly yet for handling input.  The route '/pntDevs' retrieves a json of all the devices currently in the database, and should be working correctly for Angular to use.